# api/openapi/clob.yaml
openapi: "3.0.3"
info:
  title: Poly Paper Trading - CLOB API
  version: "1.0.0"
  description: |
    Polymarket-compatible CLOB API for paper trading.
    All paths are relative to the /clob mount point.
    Bots set host to http://localhost:8080/clob.
paths:
...
  /order:
    post:
      operationId: postOrder
      tags: [trading]
      summary: Place a single order
      parameters:
        - $ref: "#/components/parameters/PolyAddress"
        - $ref: "#/components/parameters/PolySignatureL2"
        - $ref: "#/components/parameters/PolyTimestamp"
        - $ref: "#/components/parameters/PolyApiKey"
        - $ref: "#/components/parameters/PolyPassphrase"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PostOrderRequest"
      responses:
        "200":
          description: Order placed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OrderResponse"
        "400":
          description: Invalid order
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    delete:
      operationId: cancelOrder
      tags: [trading]
      summary: Cancel a single order (orderID in body)
      parameters:
        - $ref: "#/components/parameters/PolyAddress"
        - $ref: "#/components/parameters/PolySignatureL2"
        - $ref: "#/components/parameters/PolyTimestamp"
        - $ref: "#/components/parameters/PolyApiKey"
        - $ref: "#/components/parameters/PolyPassphrase"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CancelOrderRequest"
      responses:
        "200":
          description: Order canceled
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CancelOrdersResponse"
        "404":
          description: Order not found
  /orders:
    post:
      operationId: postOrders
      tags: [trading]
      summary: Place batch orders (up to 15)
      parameters:
        - $ref: "#/components/parameters/PolyAddress"
        - $ref: "#/components/parameters/PolySignatureL2"
        - $ref: "#/components/parameters/PolyTimestamp"
        - $ref: "#/components/parameters/PolyApiKey"
        - $ref: "#/components/parameters/PolyPassphrase"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: array
              items:
                $ref: "#/components/schemas/PostOrderRequest"
              maxItems: 15
      responses:
        "200":
          description: Orders placed
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/OrderResponse"
    delete:
      operationId: cancelOrders
      tags: [trading]
      summary: Cancel batch orders (array of IDs in body)
      parameters:
        - $ref: "#/components/parameters/PolyAddress"
        - $ref: "#/components/parameters/PolySignatureL2"
        - $ref: "#/components/parameters/PolyTimestamp"
        - $ref: "#/components/parameters/PolyApiKey"
        - $ref: "#/components/parameters/PolyPassphrase"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: array
              items:
                type: string
      responses:
        "200":
          description: Orders canceled
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CancelOrdersResponse"
  /cancel-all:
    delete:
      operationId: cancelAll
      tags: [trading]
      summary: Cancel all user orders
      parameters:
        - $ref: "#/components/parameters/PolyAddress"
        - $ref: "#/components/parameters/PolySignatureL2"
        - $ref: "#/components/parameters/PolyTimestamp"
        - $ref: "#/components/parameters/PolyApiKey"
        - $ref: "#/components/parameters/PolyPassphrase"
      responses:
        "200":
          description: All orders canceled
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CancelOrdersResponse"
  /cancel-market-orders:
    delete:
      operationId: cancelMarketOrders
      tags: [trading]
      summary: Cancel all orders for a specific market
      parameters:
        - $ref: "#/components/parameters/PolyAddress"
        - $ref: "#/components/parameters/PolySignatureL2"
        - $ref: "#/components/parameters/PolyTimestamp"
        - $ref: "#/components/parameters/PolyApiKey"
        - $ref: "#/components/parameters/PolyPassphrase"
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                market:
                  type: string
                asset_id:
                  type: string
      responses:
        "200":
          description: Market orders canceled
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CancelOrdersResponse"

  # --- Query (L2) ---
  /data/order/{orderId}:
    get:
      operationId: getOrder
      tags: [query]
      summary: Get order by ID
      parameters:
        - name: orderId
          in: path
          required: true
          schema:
            type: string
        - $ref: "#/components/parameters/PolyAddress"
        - $ref: "#/components/parameters/PolySignatureL2"
        - $ref: "#/components/parameters/PolyTimestamp"
        - $ref: "#/components/parameters/PolyApiKey"
        - $ref: "#/components/parameters/PolyPassphrase"
      responses:
        "200":
          description: Order details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OpenOrder"
  /data/orders:
    get:
      operationId: getOrders
      tags: [query]
      summary: Get user orders (cursor-paginated)
      parameters:
        - name: market
          in: query
          schema:
            type: string
        - name: asset_id
          in: query
          schema:
            type: string
        - name: next_cursor
          in: query
          schema:
            type: string
        - $ref: "#/components/parameters/PolyAddress"
        - $ref: "#/components/parameters/PolySignatureL2"
        - $ref: "#/components/parameters/PolyTimestamp"
        - $ref: "#/components/parameters/PolyApiKey"
        - $ref: "#/components/parameters/PolyPassphrase"
      responses:
        "200":
          description: Paginated orders
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PaginatedOrders"
  /data/trades:
    get:
      operationId: getTrades
      tags: [query]
      summary: Get user trades (cursor-paginated)
      parameters:
        - name: market
          in: query
          schema:
            type: string
        - name: asset_id
          in: query
          schema:
            type: string
        - name: next_cursor
          in: query
          schema:
            type: string
        - $ref: "#/components/parameters/PolyAddress"
        - $ref: "#/components/parameters/PolySignatureL2"
        - $ref: "#/components/parameters/PolyTimestamp"
        - $ref: "#/components/parameters/PolyApiKey"
        - $ref: "#/components/parameters/PolyPassphrase"
      responses:
        "200":
          description: Paginated trades
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PaginatedTrades"
  /balance-allowance:
    get:
      operationId: getBalanceAllowance
      tags: [query]
      summary: Get balance and allowance per asset type
      parameters:
        - name: asset_type
          in: query
          schema:
            type: string
            enum: [COLLATERAL, CONDITIONAL]
        - name: token_id
          in: query
          schema:
            type: string
        - name: signature_type
          in: query
          schema:
            type: integer
        - $ref: "#/components/parameters/PolyAddress"
        - $ref: "#/components/parameters/PolySignatureL2"
        - $ref: "#/components/parameters/PolyTimestamp"
        - $ref: "#/components/parameters/PolyApiKey"
        - $ref: "#/components/parameters/PolyPassphrase"
      responses:
        "200":
          description: Balance and allowance
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BalanceAllowanceResponse"
  /balance-allowance/update:
    get:
      operationId: updateBalanceAllowance
      tags: [query]
      summary: Update/refresh balance allowance
      parameters:
        - $ref: "#/components/parameters/PolyAddress"
        - $ref: "#/components/parameters/PolySignatureL2"
        - $ref: "#/components/parameters/PolyTimestamp"
        - $ref: "#/components/parameters/PolyApiKey"
        - $ref: "#/components/parameters/PolyPassphrase"
      responses:
        "200":
          description: Balance allowance updated
  /notifications:
    get:
      operationId: getNotifications
      tags: [query]
      summary: Get user notifications
      parameters:
        - $ref: "#/components/parameters/PolyAddress"
        - $ref: "#/components/parameters/PolySignatureL2"
        - $ref: "#/components/parameters/PolyTimestamp"
        - $ref: "#/components/parameters/PolyApiKey"
        - $ref: "#/components/parameters/PolyPassphrase"
      responses:
        "200":
          description: Notifications list
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
    delete:
      operationId: dropNotifications
      tags: [query]
      summary: Drop notifications
      parameters:
        - $ref: "#/components/parameters/PolyAddress"
        - $ref: "#/components/parameters/PolySignatureL2"
        - $ref: "#/components/parameters/PolyTimestamp"
        - $ref: "#/components/parameters/PolyApiKey"
        - $ref: "#/components/parameters/PolyPassphrase"
      responses:
        "200":
          description: Notifications dropped
  /order-scoring:
    get:
      operationId: isOrderScoring
      tags: [query]
      summary: Check if order is scoring
      parameters:
        - name: orderId
          in: query
          required: true
          schema:
            type: string
        - $ref: "#/components/parameters/PolyAddress"
        - $ref: "#/components/parameters/PolySignatureL2"
        - $ref: "#/components/parameters/PolyTimestamp"
        - $ref: "#/components/parameters/PolyApiKey"
        - $ref: "#/components/parameters/PolyPassphrase"
      responses:
        "200":
          description: Scoring status
          content:
            application/json:
              schema:
                type: object
                properties:
                  scoring:
                    type: boolean
  /orders-scoring:
    post:
      operationId: areOrdersScoring
      tags: [query]
      summary: Check if orders are scoring
      parameters:
        - $ref: "#/components/parameters/PolyAddress"
        - $ref: "#/components/parameters/PolySignatureL2"
        - $ref: "#/components/parameters/PolyTimestamp"
        - $ref: "#/components/parameters/PolyApiKey"
        - $ref: "#/components/parameters/PolyPassphrase"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                orderIds:
                  type: array
                  items:
                    type: string
      responses:
        "200":
          description: Scoring statuses
  /v1/heartbeats:
    post:
      operationId: postHeartbeat
      tags: [system]
      summary: Bot heartbeat keep-alive
      parameters:
        - $ref: "#/components/parameters/PolyAddress"
        - $ref: "#/components/parameters/PolySignatureL2"
        - $ref: "#/components/parameters/PolyTimestamp"
        - $ref: "#/components/parameters/PolyApiKey"
        - $ref: "#/components/parameters/PolyPassphrase"
      responses:
        "200":
          description: Heartbeat acknowledged

  # --- Market Data (public, proxied to Polymarket) ---
...
components:
  parameters:
    PolyAddress:
      name: POLY_ADDRESS
      in: header
      required: true
      schema:
        type: string
...
    PolySignatureL2:
      name: POLY_SIGNATURE
      in: header
      required: true
      schema:
        type: string
      description: HMAC-SHA256 signature (L2)
    PolyTimestamp:
      name: POLY_TIMESTAMP
      in: header
      required: true
      schema:
        type: string
...
    PolyApiKey:
      name: POLY_API_KEY
      in: header
      required: true
      schema:
        type: string
    PolyPassphrase:
      name: POLY_PASSPHRASE
      in: header
      required: true
      schema:
        type: string
  schemas:
...
    PostOrderRequest:
      type: object
      required: [order, owner, orderType]
      properties:
        order:
          $ref: "#/components/schemas/SignedOrder"
        owner:
          type: string
        orderType:
          type: string
          enum: [GTC, FOK, GTD, FAK]
        postOnly:
          type: boolean
          default: false
    SignedOrder:
      type: object
      properties:
        salt:
          type: string
        maker:
          type: string
        signer:
          type: string
        taker:
          type: string
        tokenId:
          type: string
        makerAmount:
          type: string
        takerAmount:
          type: string
        expiration:
          type: string
        nonce:
          type: string
        feeRateBps:
          type: string
        side:
          type: string
          enum: [BUY, SELL]
        signatureType:
          type: integer
        signature:
          type: string
    OrderResponse:
      type: object
      properties:
        success:
          type: boolean
        errorMsg:
          type: string
        orderID:
          type: string
        transactionsHashes:
          type: array
          items:
            type: string
        status:
          type: string
        takingAmount:
          type: string
        makingAmount:
          type: string
    CancelOrderRequest:
      type: object
      required: [orderID]
      properties:
        orderID:
          type: string
    CancelOrdersResponse:
      type: object
      properties:
        canceled:
          type: array
          items:
            type: string
        not_canceled:
          type: object
          additionalProperties:
            type: string
    OpenOrder:
      type: object
      properties:
        id:
          type: string
        status:
          type: string
        owner:
          type: string
        maker_address:
          type: string
        market:
          type: string
        asset_id:
          type: string
        side:
          type: string
        original_size:
          type: string
        size_matched:
          type: string
        price:
          type: string
        associate_trades:
          type: array
          items:
            type: string
        outcome:
          type: string
        created_at:
          type: string
        expiration:
          type: string
        order_type:
          type: string
    TradeItem:
      type: object
      properties:
        id:
          type: string
        taker_order_id:
          type: string
        market:
          type: string
        asset_id:
          type: string
        side:
          type: string
        size:
          type: string
        fee_rate_bps:
          type: string
        price:
          type: string
        status:
          type: string
        match_time:
          type: string
        last_update:
          type: string
        outcome:
          type: string
        owner:
          type: string
        maker_address:
          type: string
        bucket_index:
          type: integer
        transaction_hash:
          type: string
        trader_side:
          type: string
        maker_orders:
          type: array
          items:
            type: object
    PaginatedOrders:
      type: object
      properties:
        next_cursor:
          type: string
        data:
          type: array
          items:
            $ref: "#/components/schemas/OpenOrder"
    PaginatedTrades:
      type: object
      properties:
        next_cursor:
          type: string
        data:
          type: array
          items:
            $ref: "#/components/schemas/TradeItem"
    BalanceAllowanceResponse:
      type: object
      properties:
        balance:
          type: string
        allowance:
          type: string
    ErrorResponse:
      type: object
      properties:
        error:
          type: string